Added my manual version, feel free to modify it.
It uses a page format where you simply add another <div> with id="page#" where # is a number.
You then go ahead and set the number of pages in the script.js file.

I have copied over most of the style from the current manual, so it works roughly the same.